real update feed

+++ features/user/delivery/handler.go
@@ -4,6 +4,8 @@ package delivery
 import (
 	"be13/project/features/user"
 	"be13/project/helper"
+	"be13/project/middlewares"
+	"log"
 	"net/http"
 	"strconv"
 
@@ -18,22 +20,19 @@ func New(Service user.ServiceEntities, e *echo.Echo) {
 	handler := &UserDeliv{
 		UserService: Service,
 	}
-	// e.GET("/user", handler.GetAll) // memanggil func getall
-	e.POST("/user", handler.Create)
-	e.GET("/user", handler.GetAll)
-	e.PUT("/user/:id", handler.Update)
-	// e.POST("/auth", handler.Login)
-	// e.PUT("/user/:id", handler.Update)
-	// e.GET("/user/:id", handler.GetById)
-	e.DELETE("/user/:id", handler.DeleteById)
+
+	e.POST("/user", handler.Create, middlewares.JWTMiddleware())
+	e.GET("/user", handler.GetAll, middlewares.JWTMiddleware())
+	e.PUT("/user/:id", handler.Update, middlewares.JWTMiddleware())
+	e.DELETE("/user/:id", handler.DeleteById, middlewares.JWTMiddleware())
 }
 func (delivery *UserDeliv) Create(c echo.Context) error {
 
-	// roletoken := middlewares.ExtractTokenUserRole(c)
-	// log.Println("Role Token", roletoken)
-	// if roletoken != "admin" {
-	// 	return c.JSON(http.StatusUnauthorized, helper.PesanGagalHelper("tidak bisa diakses khusus admin!!!"))
-	// }
+	roletoken := middlewares.ExtractTokenUserRole(c)
+	log.Println("Role Token", roletoken)
+	if roletoken != "admin" {
+		return c.JSON(http.StatusUnauthorized, helper.PesanGagalHelper("tidak bisa diakses khusus admin!!!"))
+	}
 
 	Inputuser := UserRequest{} //penangkapan data user reques dari entities user
 	errbind := c.Bind(&Inputuser)
@@ -64,11 +63,11 @@ func (delivery *UserDeliv) GetAll(c echo.Context) error {
 }
 func (delivery *UserDeliv) Update(c echo.Context) error {
 
-	// roletoken := middlewares.ExtractTokenUserRole(c)
-	// log.Println("Role Token", roletoken)
-	// if roletoken != "admin" {
-	// 	return c.JSON(http.StatusUnauthorized, helper.PesanGagalHelper("tidak bisa diakses khusus admin!!!"))
-	// }
+	roletoken := middlewares.ExtractTokenUserRole(c)
+	log.Println("Role Token", roletoken)
+	if roletoken != "admin" {
+		return c.JSON(http.StatusUnauthorized, helper.PesanGagalHelper("tidak bisa diakses khusus admin!!!"))
+	}
 	id, _ := strconv.Atoi(c.Param("id"))
 
 	userInput := UserRequest{}
@@ -85,6 +84,11 @@ func (delivery *UserDeliv) Update(c echo.Context) error {
 	return c.JSON(http.StatusCreated, helper.PesanSuksesHelper("success Update data"))
 }
 func (delivery *UserDeliv) DeleteById(c echo.Context) error {
+	roletoken := middlewares.ExtractTokenUserRole(c)
+	log.Println("Role Token", roletoken)
+	if roletoken != "admin" {
+		return c.JSON(http.StatusUnauthorized, helper.PesanGagalHelper("tidak bisa diakses khusus admin!!!"))
+	}
 	id, _ := strconv.Atoi(c.Param("id"))
 	del, err := delivery.UserService.DeleteById(id) //memanggil fungsi service yang ada di folder service
